WIP: get rid of temporary lib/command-types
