+++ lib/command.ts
@@ -1,3 +1,6 @@
+import { Dictionary } from './common-types'
+import { YargsInstance } from './yargs-types'
+
 'use strict'
 
 const inspect = require('util').inspect
@@ -433,4 +436,25 @@ module.exports = function command (yargs, usage, validation, globalMiddleware) {
   }
 
   return self
+}
+
+/** Instance of the command module. */
+export interface CommandInstance {
+  getCommandHandlers (): Dictionary<CommandHandler>
+  getCommands (): string[]
+}
+
+interface CommandHandler {
+  builder: CommandBuilder
+}
+
+// To be completed later with other CommandBuilder flavours
+type CommandBuilder = FunctionCommandBuilder
+
+interface FunctionCommandBuilder {
+  (y: YargsInstance): YargsInstance
+}
+
+export function isFunctionCommandBuilder (builder: CommandBuilder): builder is FunctionCommandBuilder {
+  return typeof builder === 'function'
 }
